OSIS-2255 move simplifiedvolume form to its own file and change prefix
