+++ base/forms/learning_unit/edition_volume.py
@@ -28,22 +28,23 @@ from collections import OrderedDict
 from django import forms
 from django.db import transaction
 from django.db.models import Prefetch
-from django.forms import formset_factory, modelformset_factory
+from django.forms import formset_factory
 from django.utils.translation import ugettext_lazy as _
 
 from base.business.learning_unit_year_with_context import ENTITY_TYPES_VOLUME
 from base.business.learning_units import edition
 from base.business.learning_units.edition import check_postponement_conflict_report_errors
 from base.forms.common import STEP_HALF_INTEGER
-from base.forms.learning_unit.learning_unit_create import DEFAULT_ACRONYM_COMPONENT
 from base.forms.utils.emptyfield import EmptyField
 from base.models.entity_component_year import EntityComponentYear
 from base.models.enums import entity_container_year_link_type as entity_types
-from base.models.enums.component_type import PRACTICAL_EXERCISES, LECTURING
-from base.models.enums.learning_container_year_types import CONTAINER_TYPE_WITH_DEFAULT_COMPONENT, \
-    LEARNING_CONTAINER_YEAR_TYPES_CANT_UPDATE_BY_FACULTY
-from base.models.learning_component_year import LearningComponentYear
+from base.models.enums.learning_container_year_types import LEARNING_CONTAINER_YEAR_TYPES_CANT_UPDATE_BY_FACULTY
 from base.models.learning_unit_component import LearningUnitComponent
+
+
+class StepHalfIntegerWidget(forms.NumberInput):
+    def __init__(self):
+        super().__init__(attrs={'step': STEP_HALF_INTEGER})
 
 
 class VolumeField(forms.DecimalField):
@@ -60,19 +61,19 @@ class VolumeEditionForm(forms.Form):
     volume_q1 = VolumeField(
         label=_('Q1'),
         help_text=_('Volume Q1'),
-        widget=forms.NumberInput(attrs={'step': STEP_HALF_INTEGER}),
+        widget=StepHalfIntegerWidget(),
     )
     add_field = EmptyField(label='+')
     volume_q2 = VolumeField(
         label=_('Q2'),
         help_text=_('Volume Q2'),
-        widget=forms.NumberInput(attrs={'step': STEP_HALF_INTEGER}),
+        widget=StepHalfIntegerWidget(),
     )
     equal_field_1 = EmptyField(label='=')
     volume_total = VolumeField(
         label=_('Vol. annual'),
         help_text=_('The annual volume must be equal to the sum of the volumes Q1 and Q2'),
-        widget=forms.NumberInput(attrs={'step': STEP_HALF_INTEGER}),
+        widget=StepHalfIntegerWidget(),
     )
     help_volume_total = "{} = {} + {}".format(_('Volume total annual'), _('Volume Q1'), _('Volume Q2'))
     closing_parenthesis_field = EmptyField(label=')')
@@ -106,7 +107,7 @@ class VolumeEditionForm(forms.Form):
             self.fields["volume_" + key.lower()] = VolumeField(
                 label=entity.acronym,
                 help_text=entity.title,
-                widget=forms.NumberInput(attrs={'step': STEP_HALF_INTEGER}),
+                widget=StepHalfIntegerWidget(),
             )
             if i != len(entities_to_add) - 1:
                 self.fields["add" + key.lower()] = EmptyField(label='+')
@@ -224,9 +225,6 @@ class VolumeEditionBaseFormset(forms.BaseFormSet):
         # Field's name must be in lowercase
         return {k.lower(): v for k, v in component_dict.items()}
 
-    def get_form_by_type(self, component_type):
-        return next(form for form in self.forms if form.component.type == component_type)
-
     def save(self, postponement):
         for form in self.forms:
             form.save(postponement)
@@ -284,169 +282,3 @@ class VolumeEditionFormsetContainer:
             for name, error in form_errors.items():
                 errors["{}-{}-{}".format(formset.prefix, i, name)] = error
         return errors
-
-
-class SimplifiedVolumeForm(forms.ModelForm):
-    _learning_unit_year = None
-    _entity_containers = []
-
-    add_field = EmptyField(label="+")
-    equal_field = EmptyField(label='=')
-
-    def __init__(self, *args, **kwargs):
-        self.component_type = kwargs.pop('component_type')
-        self.is_faculty_manager = kwargs.pop('is_faculty_manager', False)
-
-        super().__init__(*args, **kwargs)
-        self.instance.type = self.component_type
-        self.instance.acronym = DEFAULT_ACRONYM_COMPONENT[self.component_type]
-
-    class Meta:
-        model = LearningComponentYear
-        fields = (
-            'planned_classes',
-            'hourly_volume_total_annual',
-            'hourly_volume_partial_q1',
-            'hourly_volume_partial_q2'
-        )
-        widgets = {
-            'hourly_volume_total_annual': forms.NumberInput(attrs={'step': STEP_HALF_INTEGER}),
-            'hourly_volume_partial_q1': forms.NumberInput(attrs={'step': STEP_HALF_INTEGER}),
-            'hourly_volume_partial_q2': forms.NumberInput(attrs={'step': STEP_HALF_INTEGER}),
-            'volume_declared_vacant': forms.NumberInput(attrs={'step': STEP_HALF_INTEGER}),
-        }
-
-    def clean(self):
-        """
-        Prevent faculty users to a volume to 0 if there was a value other than 0.
-        Also, prevent the faculty user from putting a volume if its value was 0.
-        # FIXME Refactor this method with the clean of VolumeEditionForm
-        """
-        cleaned_data = super().clean()
-
-        if self.is_faculty_manager:
-            if 0 in [self.instance.hourly_volume_partial_q1, self.instance.hourly_volume_partial_q2]:
-                if 0 not in [self.cleaned_data.get("hourly_volume_partial_q1"),
-                             self.cleaned_data.get("hourly_volume_partial_q2")]:
-                    self.add_error("hourly_volume_partial_q1", _("One of the partial volumes must have a value to 0."))
-                    self.add_error("hourly_volume_partial_q2", _("One of the partial volumes must have a value to 0."))
-
-            else:
-                if self.cleaned_data.get("hourly_volume_partial_q1") == 0:
-                    self.add_error("hourly_volume_partial_q1", _("The volume can not be set to 0."))
-                if self.cleaned_data.get("hourly_volume_partial_q2") == 0:
-                    self.add_error("hourly_volume_partial_q2", _("The volume can not be set to 0."))
-
-        return cleaned_data
-
-    def save(self, commit=True):
-        if self.need_to_create_untyped_component():
-            self.instance.acronym = DEFAULT_ACRONYM_COMPONENT[None]
-            self.instance.type = None
-            # In case of untyped component, we just need to create only 1 component (not more)
-            if not self.is_first_form_in_formset():
-                return None
-        return self._create_structure_components(commit)
-
-    def need_to_create_untyped_component(self):
-        container_type = self._learning_unit_year.learning_container_year.container_type
-        return container_type not in CONTAINER_TYPE_WITH_DEFAULT_COMPONENT
-
-    def is_first_form_in_formset(self):
-        return self.prefix == "form-0"
-
-    def _create_structure_components(self, commit):
-        self.instance.learning_container_year = self._learning_unit_year.learning_container_year
-
-        if self.instance.hourly_volume_total_annual is None or self.instance.hourly_volume_total_annual == 0:
-            self.instance.planned_classes = 0
-        else:
-            self.instance.planned_classes = 1
-        instance = super().save(commit)
-        LearningUnitComponent.objects.update_or_create(
-            learning_unit_year=self._learning_unit_year,
-            learning_component_year=instance
-        )
-        requirement_entity_containers = self._get_requirement_entity_container()
-        for requirement_entity_container in requirement_entity_containers:
-            if not self.instance.hourly_volume_total_annual and self.initial:
-                self._get_initial_volume_data()
-            learning_unit_components = LearningUnitComponent.objects.filter(
-                learning_unit_year__learning_container_year=self._learning_unit_year.learning_container_year)
-            self._create_entity_component_years(learning_unit_components, requirement_entity_container)
-        return instance
-
-    def _create_entity_component_years(self, learning_unit_components, requirement_entity_container):
-        for learning_unit_component in learning_unit_components:
-            EntityComponentYear.objects.update_or_create(
-                entity_container_year=requirement_entity_container,
-                learning_component_year=learning_unit_component.learning_component_year
-            )
-
-    def _get_initial_volume_data(self):
-        self.instance.hourly_volume_total_annual = self.initial.get('hourly_volume_total_annual')
-        self.instance.hourly_volume_partial_q1 = self.initial.get('hourly_volume_partial_q1')
-        self.instance.hourly_volume_partial_q2 = self.initial.get('hourly_volume_partial_q2')
-        self.instance.save()
-
-    def _get_requirement_entity_container(self):
-        requirement_entity_containers = []
-        for entity_container_year in self._entity_containers:
-            if entity_container_year and entity_container_year.type != entity_types.ALLOCATION_ENTITY:
-                requirement_entity_containers.append(entity_container_year)
-        return requirement_entity_containers
-
-
-class SimplifiedVolumeFormset(forms.BaseModelFormSet):
-    def __init__(self, data, person, *args, **kwargs):
-        super().__init__(data, *args, **kwargs)
-        self.is_faculty_manager = person.is_faculty_manager() and not person.is_central_manager()
-
-    def get_form_kwargs(self, index):
-        kwargs = super().get_form_kwargs(index)
-        kwargs['component_type'] = [LECTURING, PRACTICAL_EXERCISES][index]
-        kwargs['is_faculty_manager'] = self.is_faculty_manager
-        return kwargs
-
-    @property
-    def fields(self):
-        fields = OrderedDict()
-        for form_instance in self.forms:
-            fields.update(form_instance.fields)
-        return fields
-
-    @property
-    def instances_data(self):
-        data = {}
-        zip_form_and_initial_forms = zip(self.forms, self.initial_forms)
-        for form_instance, initial_form in zip_form_and_initial_forms:
-            for col in ['hourly_volume_total_annual', 'hourly_volume_partial_q1', 'hourly_volume_partial_q2']:
-                value = getattr(form_instance.instance, col, None) or getattr(initial_form.instance, col, None)
-                data[_(form_instance.instance.type) + ' (' + self.label_fields[col].lower() + ')'] = value
-        return data
-
-    @property
-    def label_fields(self):
-        """ Return a dictionary with the label of all fields """
-        data = {}
-        for form_instance in self.forms:
-            data.update({
-                key: field.label for key, field in form_instance.fields.items()
-            })
-        return data
-
-    def save_all_forms(self, learning_unit_year, entity_container_years, commit=True):
-        for form in self.forms:
-            form._learning_unit_year = learning_unit_year
-            form._entity_containers = entity_container_years
-            form.save()
-        return super().save(commit)
-
-
-SimplifiedVolumeManagementForm = modelformset_factory(
-    model=LearningComponentYear,
-    form=SimplifiedVolumeForm,
-    formset=SimplifiedVolumeFormset,
-    extra=2,
-    max_num=2
-)
